changed some values to optioals to avoid null checks, added listening for status change of bluetooth connection, changed structure to be compatible with dependency injection over activity (roboguice), implemented roboeletric for consuming rest services refs #154, #162, #180

+++ SmartPlane/src/test/java/com/tobyrich/app/SmartPlane/dispatcher/SendDataServiceTest.java
@@ -45,8 +45,8 @@ public class SendDataServiceTest {
     @Test
     public void testSendMotorDataInteger() throws Exception {
         // Given
-        Map<Integer, Integer> map = new LinkedHashMap<Integer, Integer>();
-        map.put(5, 5);
+        Map<Long, Short> map = new LinkedHashMap<Long, Short>();
+        map.put(5L, (short) 5);
 
         // When
         AchievementService service = retrofit.create(AchievementService.class);
